# openapi_schema.yaml
openapi: 3.0.3
info:
  title: KG Query API
  version: 1.1.0
  description: >
    Knowledge graph query service for Coze. Includes entity lookup,
    path retrieval, fuzzy search, and mechanism-chain explanation.

servers:
  - url: https://kg-plugin-production.up.railway.app
    description: Production service

security:
  - ApiKeyAuth: []

components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  schemas:
    Node:
      type: object
      properties:
        id:
          type: string
        labels:
          type: array
          items:
            type: string
        properties:
          type: object
          additionalProperties: true
    Edge:
      type: object
      properties:
        id:
          type: string
        type:
          type: string
        source:
          type: string
        target:
          type: string
        properties:
          type: object
          additionalProperties: true
    Triple:
      type: object
      properties:
        start:
          $ref: '#/components/schemas/Node'
        relationship:
          type: object
          properties:
            type:
              type: string
            properties:
              type: object
              additionalProperties: true
        end:
          $ref: '#/components/schemas/Node'
    Path:
      type: object
      properties:
        hops:
          type: integer
        mode:
          type: string
        readable:
          type: string
        nodes:
          type: array
          items:
            $ref: '#/components/schemas/Node'
        edges:
          type: array
          items:
            $ref: '#/components/schemas/Edge'
        relationships:
          type: array
          items:
            type: object
            properties:
              type:
                type: string
              properties:
                type: object
                additionalProperties: true
    SupportingFact:
      type: object
      properties:
        entity:
          type: string
        source:
          type: string
        field:
          type: string
        fact:
          type: string

paths:
  /query_entity:
    post:
      operationId: queryEntity
      summary: Query entity
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [entity_name]
              properties:
                entity_name:
                  type: string
                fuzzy:
                  type: boolean
                  default: false
      responses:
        '200':
          description: Entity query result
          content:
            application/json:
              schema:
                type: object
                properties:
                  found:
                    type: boolean
                  count:
                    type: integer
                  matched_entity:
                    type: string
                  normalized_entity:
                    type: string
                  entities:
                    type: array
                    items:
                      $ref: '#/components/schemas/Node'

  /query_neighbors:
    post:
      operationId: queryNeighbors
      summary: Query multi-hop neighbors
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [entity_name]
              properties:
                entity_name:
                  type: string
                depth:
                  type: integer
                  default: 1
                  minimum: 1
                  maximum: 3
                rel_type:
                  type: string
                limit:
                  type: integer
                  default: 20
                  minimum: 1
                  maximum: 50
      responses:
        '200':
          description: Neighbor query result
          content:
            application/json:
              schema:
                type: object
                properties:
                  found:
                    type: boolean
                  count:
                    type: integer
                  path_count:
                    type: integer
                  source_entity:
                    type: string
                  matched_entity:
                    type: string
                  depth:
                    type: integer
                  paths:
                    type: array
                    items:
                      $ref: '#/components/schemas/Path'
                  triples:
                    type: array
                    items:
                      $ref: '#/components/schemas/Triple'

  /find_path:
    post:
      operationId: findPath
      summary: Find shortest path with directed-first fallback
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [start_entity, end_entity]
              properties:
                start_entity:
                  type: string
                end_entity:
                  type: string
                max_hops:
                  type: integer
                  default: 4
                  minimum: 1
                  maximum: 6
      responses:
        '200':
          description: Path query result
          content:
            application/json:
              schema:
                type: object
                properties:
                  found:
                    type: boolean
                  count:
                    type: integer
                  mode:
                    type: string
                  matched_start_entity:
                    type: string
                  matched_end_entity:
                    type: string
                  message:
                    type: string
                  paths:
                    type: array
                    items:
                      $ref: '#/components/schemas/Path'

  /fuzzy_search:
    post:
      operationId: fuzzySearch
      summary: Fuzzy search entities
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [keyword]
              properties:
                keyword:
                  type: string
                limit:
                  type: integer
                  default: 10
                  minimum: 1
                  maximum: 20
      responses:
        '200':
          description: Fuzzy search result
          content:
            application/json:
              schema:
                type: object
                properties:
                  found:
                    type: boolean
                  count:
                    type: integer
                  keyword:
                    type: string
                  normalized_keyword:
                    type: string
                  best_match:
                    type: string
                  candidates:
                    type: array
                    items:
                      $ref: '#/components/schemas/Node'

  /explain_mechanism:
    post:
      operationId: explainMechanism
      summary: Retrieve a teaching-oriented mechanism chain
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [question]
              properties:
                question:
                  type: string
                start_entity:
                  type: string
                target_entity:
                  type: string
                bridge_candidates:
                  type: array
                  items:
                    type: string
                max_hops:
                  type: integer
                  default: 5
                  minimum: 2
                  maximum: 6
                limit:
                  type: integer
                  default: 30
                  minimum: 5
                  maximum: 50
      responses:
        '200':
          description: Mechanism explanation result
          content:
            application/json:
              schema:
                type: object
                properties:
                  found:
                    type: boolean
                  question_type:
                    type: string
                  start_entity:
                    type: string
                  target_entity:
                    type: string
                  focus_entities:
                    type: array
                    items:
                      type: string
                  core_entities:
                    type: array
                    items:
                      type: string
                  resolved_bridges:
                    type: array
                    items:
                      type: string
                  missing_bridges:
                    type: array
                    items:
                      type: string
                  recommended_next_entity:
                    type: string
                  path_complete:
                    type: boolean
                  retrieval_query:
                    type: string
                  current_mechanism_chain:
                    type: string
                  answer_outline:
                    type: array
                    items:
                      type: string
                  evidence_chains:
                    type: array
                    items:
                      $ref: '#/components/schemas/Path'
                  supporting_facts:
                    type: array
                    items:
                      $ref: '#/components/schemas/SupportingFact'

  /dispatch:
    post:
      operationId: dispatch
      summary: Unified Coze dispatch endpoint
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [action]
              properties:
                action:
                  type: string
                  description: queryEntity | queryNeighbors | findPath | fuzzySearch | explainMechanism
                entity_name:
                  type: string
                fuzzy:
                  type: boolean
                  default: false
                depth:
                  type: integer
                  default: 1
                rel_type:
                  type: string
                limit:
                  type: integer
                  default: 20
                start_entity:
                  type: string
                end_entity:
                  type: string
                max_hops:
                  type: integer
                  default: 4
                keyword:
                  type: string
                question:
                  type: string
                bridge_candidates:
                  type: array
                  items:
                    type: string
      responses:
        '200':
          description: Routed result with JSON string mirror
          content:
            application/json:
              schema:
                type: object
                properties:
                  result_json:
                    type: string
                    description: Full response serialized as JSON for Coze code nodes

  /health:
    get:
      operationId: healthCheck
      summary: Health check
      security: []
      responses:
        '200':
          description: Service and Neo4j status
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  neo4j:
                    type: string
                  timestamp:
                    type: string